Use a MongoDB replica set.
I used a Kubernetes headless service to allow clients to communicate
with the replica set using the DNS name "mongo" or "mongodb". A
headless service points directly to the backing VMs/Pods instead of
providing a load balancer for them. When a mongo client connects to the
service, the connection is forwarded to one of the nodes in the replica
set. The client should then be able to discover which node is the
primary and read/write to it.