Use CloudFront to cache static assets.
Cached static assets are now served from CloudFront and we no longer
use an S3 bucket. If an asset hasn't yet been cached by CloudFront, it
is served from an edxapp pod and then cached.

Pods assume that static assets can be obtained from a mounted
volume. Currently, this volume is a directory that lives on the host.
If and when Kubernetes supports a generic shared volume that is much
less complex than nfs or glusterfs, we should consider using that
instead.

Although static assets are retrieved from a volume, it still seems to
be necessary to run `manage.py collectstatic` when a pod starts,
otherwise URLs are not constructed properly. For this reason, we use
NonPackagingPipelineCachedStorage as the storage backend because it
allows us to skip packaging and makes collectstatic relatively fast.